requirements.txt
SpeechRecognition==3.10.0
PyAudio==0.2.14; python_version<"3.14"
pyttsx3==2.90
python-dotenv==1.0.0
flask==3.0.0
flask-cors>=4.0.0
psutil>=5.9.0
GPUtil>=1.4.0
pydub>=0.25.1
opencv-python>=4.8.0
tinytuya>=1.12.0
Pillow>=10.0.0
python-telegram-bot>=20.0
groq>=0.4.0
edge-tts>=6.1.0
pygame>=2.6.0; python_version<"3.14"
requests>=2.31.0
